avoid warning about loss of data
we use stl containers but still have the int limit
for sizes like Qt 5.x

+++ src/buffer/katetextblock.h
@@ -89,7 +89,7 @@ public:
      */
     int lines() const
     {
-        return m_lines.size();
+        return static_cast<int>(m_lines.size());
     }
 
     /**
